Testing updates for modified F-test interface
Includes pulling updated binary test data for vectorstats.
Includes some minor fixes to said interface change.

+++ src/stats/permtest.cpp
@@ -274,7 +274,7 @@ namespace MR
                              matrix_type& uncorrected_pvalues)
       {
         assert (stats_calculator);
-        assert (stats_calculator->num_elements() == mask.size());
+        assert (stats_calculator->num_elements() == size_t(mask.size()));
         Math::Stats::Shuffler shuffler (stats_calculator->num_inputs(), false, "Running permutations");
         null_dist.resize (shuffler.size(), fwe_strong ? 1 : stats_calculator->num_hypotheses());
         null_dist_contributions = count_matrix_type::Zero (stats_calculator->num_elements(), stats_calculator->num_hypotheses());
